+++ crates/types/src/archive.rs
@@ -1,20 +1,13 @@
 use core::fmt::{Display, Formatter};
 
+use alloc::vec::Vec;
+
 use crate::TinyWasmModule;
-use rkyv::{
-    access,
-    api::high::to_bytes_in_with_alloc,
-    deserialize,
-    ser::{allocator::Arena, WriterExt},
-    Archived,
-};
 
 const TWASM_MAGIC_PREFIX: &[u8; 4] = b"TWAS";
-const TWASM_VERSION: &[u8; 2] = b"02";
+const TWASM_VERSION: &[u8; 2] = b"03";
 #[rustfmt::skip]
 const TWASM_MAGIC: [u8; 16] = [ TWASM_MAGIC_PREFIX[0], TWASM_MAGIC_PREFIX[1], TWASM_MAGIC_PREFIX[2], TWASM_MAGIC_PREFIX[3], TWASM_VERSION[0], TWASM_VERSION[1], 0, 0, 0, 0, 0, 0, 0, 0, 0, 0];
-
-pub use rkyv::util::AlignedVec;
 
 fn validate_magic(wasm: &[u8]) -> Result<usize, TwasmError> {
     if wasm.len() < TWASM_MAGIC.len() || &wasm[..TWASM_MAGIC_PREFIX.len()] != TWASM_MAGIC_PREFIX {
@@ -35,7 +28,7 @@ pub enum TwasmError {
     InvalidMagic,
     InvalidVersion,
     InvalidPadding,
-    InvalidArchive(rkyv::rancor::Error),
+    InvalidArchive(postcard::Error),
 }
 
 impl Display for TwasmError {
@@ -58,20 +51,14 @@ impl TinyWasmModule {
     /// Creates a `TinyWasmModule` from a slice of bytes.
     pub fn from_twasm(wasm: &[u8]) -> Result<TinyWasmModule, TwasmError> {
         let len = validate_magic(wasm)?;
-        let root = access::<Archived<Self>, rkyv::rancor::Error>(&wasm[len..]).map_err(TwasmError::InvalidArchive)?;
-        deserialize::<TinyWasmModule, rkyv::rancor::Error>(root).map_err(TwasmError::InvalidArchive)
+
+        postcard::from_bytes(&wasm[len..]).map_err(|e| TwasmError::InvalidArchive(e))
     }
 
     /// Serializes the `TinyWasmModule` into a vector of bytes.
-    /// `AlignedVec` can be deferenced as a slice of bytes and
-    /// implements `io::Write` when the `std` feature is enabled.
-    pub fn serialize_twasm(&self) -> AlignedVec {
-        let mut arena = Arena::new();
-        let mut bytes = AlignedVec::new();
-        <AlignedVec as WriterExt<rkyv::rancor::Error>>::pad(&mut bytes, TWASM_MAGIC.len()).unwrap();
-        let mut bytes = to_bytes_in_with_alloc::<_, _, rkyv::rancor::Error>(self, bytes, arena.acquire()).unwrap();
-        bytes[..TWASM_MAGIC.len()].copy_from_slice(&TWASM_MAGIC);
-        bytes
+    pub fn serialize_twasm(&self) -> Result<Vec<u8>, TwasmError> {
+        let buf = Vec::from(TWASM_MAGIC);
+        postcard::to_extend(self, buf).map_err(|e| TwasmError::InvalidArchive(e))
     }
 }
 
@@ -82,7 +69,7 @@ mod tests {
     #[test]
     fn test_serialize() {
         let wasm = TinyWasmModule::default();
-        let twasm = wasm.serialize_twasm();
+        let twasm = wasm.serialize_twasm().expect("should serialize");
         let wasm2 = TinyWasmModule::from_twasm(&twasm).unwrap();
         assert_eq!(wasm, wasm2);
     }
@@ -90,7 +77,7 @@ mod tests {
     #[test]
     fn test_invalid_magic() {
         let wasm = TinyWasmModule::default();
-        let mut twasm = wasm.serialize_twasm();
+        let mut twasm = wasm.serialize_twasm().expect("should serialize");
         twasm[0] = 0;
         assert!(matches!(TinyWasmModule::from_twasm(&twasm), Err(TwasmError::InvalidMagic)));
     }
@@ -98,7 +85,7 @@ mod tests {
     #[test]
     fn test_invalid_version() {
         let wasm = TinyWasmModule::default();
-        let mut twasm = wasm.serialize_twasm();
+        let mut twasm = wasm.serialize_twasm().expect("should serialize");
         twasm[4] = 0;
         assert!(matches!(TinyWasmModule::from_twasm(&twasm), Err(TwasmError::InvalidVersion)));
     }